Implemented get tile method for finding out on what tile the player is standing Implemented AABB util class Fixed bug in calculation in ChunkUtils Added Air tile to tiles Implemented to String method for tiles to return tile name

+++ src/com/microrpg/world/Position.java
@@ -22,11 +22,11 @@ public class Position {
         this.x = x;
     }
 
-    public float getX() {
+    public int getX() {
         return x;
     }
 
-    public float getY() {
+    public int getY() {
         return y;
     }
